upd: rename/reorg release sub-pkg
